Added compatibility with the 3B+, 3A+, and Zero W
The Raspberry Pi 3B+, 3A+, and Zero W contain the same bluetooth module
as the original Raspberry Pi 3, which means they use the same serial port
configuration. The Sleepy-Pi-Setup.sh script did not account for this
and would incorrectly set up the serial port used to program the Sleepy Pi 2.

The Sleepy-Pi-Setup.sh script was modified to configure
the ttyS0 serial port on these boards, instead of the ttyAMA0 serial
port.